Rename colour registry
