+++ src/vs/workbench/contrib/preferences/common/settingsEditorColorRegistry.ts
